refactor: extract useTurntablePlayState hook
